requirements.txt
Django==6.0.1
djangorestframework==3.16.1
pillow==12.1.0
psycopg2==2.9.11
python-dotenv==1.2.1
django-filter==25.2
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.14
stripe==14.3.0
celery==5.6.2
django-celery-beat==2.8.1
python-dateutil==2.9.0.post0